Add failing test case for partition copy/paste
